Decompile func_8000D384_DF84

+++ include/functions.us.h
@@ -70,7 +70,7 @@ s32 func_8000D71C_E31C(Unk8007F878_404 *arg0, Unk8007F878_404 *arg1, AnimChannel
 void func_8000CD54_D954(void *arg0, AnimChannelState *arg1, u8 arg2);
 s32 func_8000CDFC_D9FC(Unk8007F878_404 *arg0, AnimChannelState *arg1, s32 arg2, s32 arg3, s32 arg4);
 void func_8000D278_DE78(Unk8007F878_404 *arg0, AnimChannelState *arg1, s32 arg2, s32 arg3);
-void func_8000D384_DF84(AnimFrameData *arg0, AnimFrameData *arg1, s32 arg2, AnimFrameData *arg3);
+void func_8000D384_DF84(AnimFrameData *arg0, AnimFrameData *arg1, f32 arg2, AnimFrameData *arg3);
 void func_8000D588_E188(Unk8007F878_404 *arg0, Unk8007F878_404 *arg1, AnimChannelState *arg2, u8 arg3, f32 arg4);
 void func_8000DAFC_E6FC(Unk8007F878_404 *arg0, Unk8007F878_404 *arg1, AnimChannelState *arg2, s32 arg3, s32 arg4, f32 arg5);
 void func_8000DC9C_E89C(s32, s32);
